requirements.txt
accelerate
addict
diffusers
gradio==3.36.1
huggingface_hub
matplotlib
numpy
onnxruntime
opencv_python
Pillow
pycocotools
PyYAML
requests
setuptools
supervision
termcolor
timm
torch
torchvision
transformers
yapf
numba
scipy 
safetensors
pynvml
sentencepiece

lama-cleaner==1.1.2
openmim==0.1.5
mmcv==2.0.0
